pyproject.toml
[project]
name = "django-forum"
version = "0.1.1"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-allauth>=65.7.0",
    "django-auditlog>=3.4.1",
    "django-constance[database]>=4.3.5",
    "django-environ>=0.13.0",
    "django-log-viewer>=1.1.8",
    "django-lucus==0.1.8",
    "django-ratelimit>=4.1.0",
    "django-simple-captcha>=0.6.3",
    "django-solo>=2.5.1",
    "django-split-settings>=1.3.2",
    "django-summernote>=0.8.20.0",
    "asgiref>=3.11.1",
    "bleach>=6.3.0",
    "gunicorn>=25.3.0",
    "markdown>=3.8",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.0",
    "python-dateutil>=2.9.0.post0",
    "redis>=5.0.0",
    "six>=1.17.0",
    "sqlparse>=0.5.5",
    "tinycss2>=1.5.1",
    "webencodings>=0.5.1",
]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = ["E", "F", "UP"]

[tool.ruff.lint.per-file-ignores]
# Settings/components are included via split-settings and expect names
# (e.g. DEBUG, LOGS_DIR, MIDDLEWARE) to exist in the including module.
"core/settings/components/**/*.py" = ["F821"]

# These settings modules intentionally use star-import to re-export base settings.
"core/settings/development.py" = ["F403", "F405"]
"core/settings/production.py" = ["F403", "F405"]
